requirements.txt
dnspython
numpy
selenium